Reorganized the repo and move the T2B reader into its own file Closes #1, closes #2 : This makes the project more logical and more ready for expansion
