+++ .gitignore
@@ -1,3 +1,4 @@
 /target
 *.cfg.bin
 *.json
+/test_files
