+++ ievr_cfg_bin_editor-core/Cargo.toml
@@ -4,6 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-crc-fast = "1.10.0"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
+crc-fast = { version = "1.10", default-features = false, features = ["std"] }
+serde = { version = "1", default-features = false, features = ["derive", "std"] }
+serde_json = { version = "1", default-features = false, features = ["std"] }
